fix(kv): avoid quota-heavy cache maintenance
